Fix gradlew deps and cacert to be able to run game
The Adventure Editor uses gradle to run the game and gradle downloads
deps from the network. To do this libzip is needed and we should add the
env JAVA_OPTS to the correct cacerts to let gradlew download deps.

I've also updated the openjdk sdk extension to use the lastest one.